+++ packages/openrouter/src/client.ts
@@ -1,3 +1,4 @@
+//TODO: Once everything is fully migrated to the mastra package, we can remove this internal package
 import { createOpenRouter } from "@openrouter/ai-sdk-provider";
 
 export function createOpenrouterClient(apiKey: string) {