+++ packages/tavily/src/client.ts
@@ -1,3 +1,5 @@
+//TODO: Once everything is fully migrated to the mastra package, we can remove this internal package
+
 import { tavily } from "@tavily/core";
 
 export const createTavilyClient = (apiKey: string) => {